[C++] Warnings tidy up.

+++ aeron-client/src/main/cpp/concurrent/status/LocalSocketAddressStatus.h
@@ -59,7 +59,7 @@ public:
                     std::uint8_t *keyPtr = countersReader.metaDataBuffer().buffer() +
                         static_cast<std::size_t>(countersReader.metadataOffset(i)) +
                         static_cast<std::size_t>(CountersReader::KEY_OFFSET);
-                    const AtomicBuffer keyBuffer(keyPtr, CountersReader::MAX_KEY_LENGTH);
+                    const AtomicBuffer keyBuffer(keyPtr, static_cast<std::size_t>(CountersReader::MAX_KEY_LENGTH));
 
                     if (countersReader.getCounterTypeId(i) == LOCAL_SOCKET_ADDRESS_STATUS_TYPE_ID &&
                         channelStatusId == channelStatusIdFromKeyBuffer(keyBuffer) &&
